+++ src/sw/allotmentclub/browser/tests/test_account.py
@@ -90,3 +90,30 @@ def test_SEPASammlerUpdateView_2(browser):
         'http://localhost/accounts/sepa_sammler/{}/update'.format(sammler.id))
     assert len(SEPASammlerEntry.query().all()) == 2
     SEPASammlerEntry.query().filter(SEPASammlerEntry.value == 821300).one()
+
+
+def test_SEPASammlerExportView_1(browser):
+    """Testing SEPASammler Energieabschlag 1 XML export."""
+    from sw.allotmentclub import SEPASammler, SEPASammlerEntry, BookingKind
+    from sw.allotmentclub import Member
+    import datetime
+    import lxml.etree
+    kind = BookingKind.find_or_create(
+        title='Energieabschlag I', shorttitle='ENA1')
+    sammler = SEPASammler.create(
+        booking_day='2018-03-31', accounting_year=2018, kind=kind)
+    for iban, value in (('DE12500105170648489890', 2540500),
+                        ('EE342200221034126658', 8213400)):
+        SEPASammlerEntry.find_or_create(
+            sepasammler=sammler, value=value,
+            member=Member.create(
+                lastname='Müller', direct_debit=True, iban=iban, bic='NOLADE',
+                direct_debit_date=datetime.date(2017, 1, 1)))
+    setUp()
+    browser.login()
+    browser.open(
+        'http://localhost/accounts/sepa_sammler/{}/export'.format(sammler.id))
+    doc = lxml.etree.fromstring(browser.contents.encode('utf-8'))
+    assert '1075.39' == doc.find('.//CtrlSum', namespaces=doc.nsmap).text
+    assert 'Müller, ' == doc.findall('.//Nm', namespaces=doc.nsmap)[-1].text
+    assert 'Mueller-' == doc.find('.//MndtId', namespaces=doc.nsmap).text